api/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
Pillow==10.2.0
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.1.0
channels-redis==4.2.0
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
anthropic>=0.40.0
